+++ src/compute-engine/solve.ts
@@ -4,7 +4,7 @@ import {
 } from './boxed-expression/public';
 import { isInequality } from './boxed-expression/utils';
 import { Rule } from './public';
-import { boxRules, matchRules } from './rules';
+import { matchRules } from './rules';
 import { expand } from './symbolic/expand';
 
 // https://en.wikipedia.org/wiki/Equation_solving
@@ -29,6 +29,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
     match: ['Multiply', '_x', '__a'],
     replace: 0,
     id: 'ax',
+    exact: false,
     condition: ({ __a }) => !__a.has('_x'),
   },
 
@@ -36,6 +37,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
   {
     match: ['Add', ['Divide', '_a', '_x'], '__b'],
     replace: Infinity,
+    exact: false,
     condition: ({ _a, __b }) => !_a.has('_x') && !__b.has('_x'),
   },
 
@@ -43,6 +45,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
   {
     match: ['Add', ['Multiply', '_x', '__a'], '__b'],
     replace: ['Divide', ['Negate', '__b'], '__a'],
+    exact: false,
     condition: ({ __a, __b }) => !__a.has('_x') && !__b.has('_x'),
   },
 
@@ -55,6 +58,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
       ['Divide', 1, '_n'],
     ],
 
+    exact: false,
     condition: ({ _a, __b, _n }) =>
       !_a.has('_x') && !__b.has('_x') && !_n.isZero,
   },
@@ -83,6 +87,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
       ],
       ['Multiply', 2, '__a'],
     ],
+    exact: false,
     condition: ({ __a, __b, __c }) =>
       !__a.has('_x') && !__b.has('_x') && !__c.has('_x'),
   },
@@ -106,6 +111,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
       ],
       ['Multiply', 2, '__a'],
     ],
+    exact: false,
     condition: ({ __a, __b, __c }) =>
       !__a.has('_x') && !__b.has('_x') && !__c.has('_x'),
   },
@@ -118,6 +124,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
       '__c',
     ],
     replace: ['Divide', ['Ln', ['Negate', ['Divide', '__c', '__a']]], '__b'],
+    exact: false,
     condition: ({ __a, __c }, ce) =>
       ((!__a.isZero && ce.div(__c, __a).isNegative) ?? false) &&
       !__a.has('_x') &&
@@ -128,6 +135,7 @@ export const UNIVARIATE_ROOTS: Rule[] = [
   {
     match: ['Add', ['Multiply', '__a', ['Exp', '_x']], '__c'],
     replace: ['Ln', ['Negate', ['Divide', '__c', '__a']]],
+    exact: false,
     condition: ({ __a, __c }, ce) =>
       ((!__a.isZero && ce.div(__c, __a).isNegative) ?? false) &&
       !__a.has('_x') &&
@@ -138,28 +146,32 @@ export const UNIVARIATE_ROOTS: Rule[] = [
   // {
   //   match: ['Add', ['Exp', '_x'], '__c'],
   //   replace: ['Ln', ['Negate', '__c']],
-  //   condition: ({ __c }) => __c.isNegative ?? false,
+  //   exact: false,
+  //  condition: ({ __c }) => __c.isNegative ?? false,
   // },
 
   // // e^(bx) + c = 0
   // {
   //   match: ['Add', ['Exp', ['Multiply', '__b', '_x']], '__c'],
   //   replace: ['Divide', ['Ln', ['Negate', '__c']], '__b'],
-  //   condition: ({ __c }) => __c.isNegative ?? false,
+  //   exact: false,
+  //  condition: ({ __c }) => __c.isNegative ?? false,
   // },
 
   // // a * log_b(x) + c = 0
   // {
   //   match: ['Add', ['Multiply', '__a', ['Log', '_x', '__b']], '__c'],
   //   replace: ['Power', '__b', ['Negate', ['Divide', '__c', '__a']]],
-  //   condition: ({ __a, __b }) => (!__a.isZero && __b.isPositive) ?? false,
+  //   exact: false,
+  //  condition: ({ __a, __b }) => (!__a.isZero && __b.isPositive) ?? false,
   // },
 
   // // a * log_b(x) = 0
   // {
   //   match: ['Multiply', '__a', ['Log', '_x', '__b']],
   //   replace: ['Power', '__b', ['Negate', ['Divide', '__c', '__a']]],
-  //   condition: ({ __a, __b }) => (!__a.isZero && __b.isPositive) ?? false,
+  //   exact: false,
+  //  condition: ({ __a, __b }) => (!__a.isZero && __b.isPositive) ?? false,
   // },
 
   // // |ax + b| + c = 0
@@ -189,9 +201,8 @@ export function findUnivariateRoots(
     expr = ce.add(expr.op1.canonical, ce.neg(expr.op2.canonical)).simplify();
   }
 
-  const rules = ce.cache('univariate-roots-rules', () =>
-    boxRules(ce, UNIVARIATE_ROOTS)
-  );
+  const rules = ce.getRuleSet('solve-univariate')!;
+
   // Make the unknown '_x' so that we can match against it
   let exprs = [expr.subs({ [x]: '_x' }, { canonical: false })];
 
@@ -356,9 +367,7 @@ export const HARMONIZATION_RULES: Rule[] = [
  */
 function harmonize(expr: BoxedExpression): BoxedExpression[] {
   const ce = expr.engine;
-  const rules = ce.cache('univariate-roots-rules', () =>
-    boxRules(ce, HARMONIZATION_RULES)
-  );
+  const rules = ce.getRuleSet('harmonization')!;
   let result = matchRules(expr, rules, { _x: ce.symbol('_x') });
 
   return result;